Add of static_casts to suppress warnings

+++ src/imu/imu_sensor.cpp
@@ -81,11 +81,11 @@ auto InertialMeasurementSensor::SetSensorValues(std::int16_t x, std::int16_t y, 
 }
 
 auto InertialMeasurementSensor::SetBit(std::uint8_t byte, std::uint8_t bit_number_between_0_and_7) noexcept -> std::uint8_t {
-  return byte | 1 << bit_number_between_0_and_7;
+  return static_cast<std::uint8_t>(byte | 1 << bit_number_between_0_and_7);
 }
 
 auto InertialMeasurementSensor::ClearBit(std::uint8_t byte, std::uint8_t bit_number_between_0_and_7) noexcept -> std::uint8_t {
-  return byte & ~SetBit(0, bit_number_between_0_and_7);
+  return static_cast<std::uint8_t>(byte & ~SetBit(0, bit_number_between_0_and_7));
 }
 
 }  // namespace imu
